overaul pricing section
