+++ clients/apps/web/src/components/Products/ProductForm/Pricing/ProductPriceMeteredUnitItem.tsx
@@ -1,0 +1,189 @@
+'use client'
+
+import CreateMeterModalContent from '@/components/Meter/CreateMeterModalContent'
+import MeterSelector from '@/components/Meter/MeterSelector'
+import { InlineModal } from '@/components/Modal/InlineModal'
+import { useModal } from '@/components/Modal/useModal'
+import { SpinnerNoMargin } from '@/components/Shared/Spinner'
+import { useMeters } from '@/hooks/queries/meters'
+import { schemas } from '@polar-sh/client'
+import Button from '@polar-sh/ui/components/atoms/Button'
+import MoneyInput from '@polar-sh/ui/components/atoms/MoneyInput'
+import {
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from '@polar-sh/ui/components/ui/form'
+import { PlusIcon } from 'lucide-react'
+import React, { useCallback } from 'react'
+import { useFormContext } from 'react-hook-form'
+import { ProductFormType } from '../ProductForm'
+import UnitAmountInput from '../UnitAmountInput'
+
+export interface ProductPriceMeteredUnitItemProps {
+  organization: schemas['Organization']
+  index: number
+  currency: string
+}
+
+export const ProductPriceMeteredUnitItem: React.FC<
+  ProductPriceMeteredUnitItemProps
+> = ({ organization, index, currency }) => {
+  const { control, setValue } = useFormContext<ProductFormType>()
+
+  const { data: meters } = useMeters(organization.id, {
+    sorting: ['name'],
+    limit: 30,
+    is_archived: false,
+  })
+
+  const {
+    isShown: isCreateMeterModalShown,
+    show: showCreateMeterModal,
+    hide: hideCreateMeterModal,
+  } = useModal(false)
+
+  const onSelectMeter = useCallback(
+    (meter: schemas['Meter']) => {
+      setValue(`prices.${index}.meter_id`, meter.id)
+      setValue(`prices.${index}.id`, '')
+    },
+    [setValue, index],
+  )
+
+  if (!meters) {
+    return (
+      <div className="flex w-full items-center justify-center py-4">
+        <SpinnerNoMargin />
+      </div>
+    )
+  }
+
+  return (
+    <>
+      {meters.items.length === 0 ? (
+        <Button
+          onClick={(e) => {
+            e.preventDefault()
+            showCreateMeterModal()
+          }}
+          size="sm"
+        >
+          Create a Meter
+        </Button>
+      ) : (
+        <>
+          <FormField
+            control={control}
+            name={`prices.${index}.meter_id`}
+            rules={{
+              required: 'This field is required',
+            }}
+            render={({ field }) => {
+              return (
+                <FormItem>
+                  <div className="flex flex-row items-center justify-between gap-x-2">
+                    <FormLabel>Meter</FormLabel>
+                    <button
+                      type="button"
+                      className="flex flex-row items-center gap-x-1 text-sm font-medium text-gray-500"
+                      onClick={(e) => {
+                        e.preventDefault()
+                        showCreateMeterModal()
+                      }}
+                    >
+                      <PlusIcon className="h-4 w-4" />
+                      Add Meter
+                    </button>
+                  </div>
+                  <FormControl>
+                    <MeterSelector
+                      organizationId={organization.id}
+                      value={field.value || null}
+                      onChange={(meterId) => {
+                        field.onChange(meterId ?? '')
+                        setValue(`prices.${index}.id`, '')
+                      }}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )
+            }}
+          />
+          <FormField
+            control={control}
+            name={`prices.${index}.unit_amount`}
+            rules={{
+              min: 0,
+              required: 'This field is required',
+            }}
+            render={({ field }) => {
+              return (
+                <FormItem>
+                  <FormLabel>Amount per unit</FormLabel>
+                  <FormControl>
+                    <UnitAmountInput
+                      {...field}
+                      name={field.name}
+                      currency={currency}
+                      value={field.value}
+                      onValueChange={(v) => {
+                        field.onChange(v)
+                        setValue(`prices.${index}.id`, '')
+                      }}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )
+            }}
+          />
+          <FormField
+            control={control}
+            name={`prices.${index}.cap_amount`}
+            render={({ field }) => {
+              return (
+                <FormItem>
+                  <FormLabel>Cap amount</FormLabel>
+                  <FormControl>
+                    <MoneyInput
+                      {...field}
+                      name={field.name}
+                      currency={currency}
+                      value={field.value}
+                      onChange={(v) => {
+                        field.onChange(v)
+                        setValue(`prices.${index}.id`, '')
+                      }}
+                      placeholder={10000}
+                    />
+                  </FormControl>
+                  <FormDescription>
+                    Optional maximum amount that can be charged, regardless of
+                    the number of units consumed.
+                  </FormDescription>
+                  <FormMessage />
+                </FormItem>
+              )
+            }}
+          />
+        </>
+      )}
+      <InlineModal
+        isShown={isCreateMeterModalShown}
+        hide={hideCreateMeterModal}
+        modalContent={
+          <CreateMeterModalContent
+            organization={organization}
+            onSelectMeter={onSelectMeter}
+            hideModal={hideCreateMeterModal}
+          />
+        }
+      />
+    </>
+  )
+}
